requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
blinker==1.6.2
    # via flask
click==8.1.5
    # via flask
flask==2.3.2
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
python-dotenv==1.0.0
    # via -r requirements.in
werkzeug==2.3.6
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
